app/requirements.txt
absl-py
astunparse
blinker
certifi
charset-normalizer
click
Cython
filelock
Flask
flatbuffers
fsspec
gast
google-pasta
grpcio
h5py
idna
imutils
itsdangerous
Jinja2
keras
Keras-Preprocessing
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
mpmath
mtcnn
namex
networkx
numpy
opencv-contrib-python
opencv-python
opt-einsum
optree
packaging
Pillow
protobuf
Pygments
requests
rich
setuptools
six
sympy
tensorboard
tensorboard-data-server
tensorflow
termcolor
torch
torchvision
typing_extensions
urllib3
Werkzeug
wheel
wrapt
